refactor(portal): extract config-lint engine into testable module
Lift the lint engine — the lenient YAML parser (parseYaml/parseValue),
the LINT_RULES check set, and a new lintConfig(yamlText) wrapper around
the parse->run-rules->severity-sort flow — out of config-lint.jsx
(458->190 LOC) into config-lint/lint.js (PR-portal-20). The rule check()
functions hold the real lint logic and were 0%-covered; 7 tests now drive
the real rules end-to-end (threshold-too-high, missing-critical-pair pair
detection, tenant counting, severity ordering). LINT_RULES is re-exported
and imported back for the rule-list render. The i18n `t` fallback moved
with the cluster (rule categories evaluate t() at module load).

config-lint.jsx legitimately shrinks 458->190 LOC (~46%), tripping the
>50% sed-damage-guard; added to .sed-damage-allowlist (it is a real
refactor, file has EOF newline + no NUL bytes).

Behavior-preserving: logic moved verbatim, useMemo now calls lintConfig;
dist rebuilt, suite 240 -> 247 tests green.

Refs: TRK-230
Self-Review-Pass-2: 4-lens + refute; real LINT_RULES tested e2e; gates green
